backend/requirements.txt
Django
djangorestframework
mysql-connector-python
mysqlclient
django-cors-headers

Pillow